Adds event emission to the message services (#38)
* Adds events to the message services

Ports events from iovns to this repository

* Remove duplicated events attributes when emitting them

The Action parameter is filled automatically so it is not required to specify it. Sender field seems to behave the same way but we need further confirmation

* Remove the sender attributes from event

We have figured out that the  field is already included by the SDK

* Fix formatting issue

* Add registerer and resources (JSON format) for register_account event

We needed to serialize the resources array into a string to fit the string value expected by the Cosmos SDK API in the event attribute. JSON is ubiquitous and is better than some custom encoding.

Co-authored-by: Dave Puchyr <dave@slim.avaritia.com>