Keep it simple
Note that the trimpaths on `-gcflags` and `-asmflags` that were previously applied should be suspects if build signatures across machines don't match.

If the faucet or block-metrics needs to be gitian built then just do `( cd cmd/faucet && make install)` after starnamed's `make install`.

Being clever with LIBRARY_PATH was futile, likely due to the number of backslashes required to escape the `!`s in the wasmvm path when getting
processed through yaml, etc.